Minor fixes from code review

+++ java/org/apache/catalina/valves/PersistentValve.java
@@ -50,8 +50,8 @@ import org.apache.juli.logging.LogFactory;
  * <p>
  * To avoid conflicts and/or errors when updating the session store, each session must only be accessed by no more than
  * one concurrent request. The {@code filter} field can be used to define requests (e.g. those for static resources)
- * that do not need access to the session and can Requests for resources that do not need to access the session and can
- * bypass the session load/save functionality provided by this Valve.
+ * that do not need access to the session and can bypass the session load/save functionality
+ * provided by this Valve.
  * <p>
  * The Valve uses a per session {@code Semaphore} to ensure that each session is accessed by no more than one request at
  * a time within a single Tomcat instance. The behaviour if multiple requests try to access the session concurrently can
@@ -332,7 +332,7 @@ public class PersistentValve extends ValveBase {
         if (session != null) {
             int maxInactiveInterval = session.getMaxInactiveInterval();
             if (maxInactiveInterval > 0) {
-                int timeIdle = (int) (session.getIdleTimeInternal() / 1000L);
+                long timeIdle = session.getIdleTimeInternal() / 1000L;
                 return timeIdle >= maxInactiveInterval;
             }
         }
